+++ CHANGELOG.md
@@ -498,3 +498,4 @@
 - 2026-04-13: Integ bettercontact, Cargo, site config, task, +4
 - 2026-04-14: Valid PORTING, send, discover, conf, +3
 - 2026-04-15: Docs task, leads, Makefile, icp, +1
+- 2026-04-16: Health campaign, users, contacts, summaries, +3